default/requirements.txt
flask[async]
wtforms
gunicorn
python-leetcode
google-cloud-secret-manager
appengine-python-standard
